# Batched write barrier for byrefs