--- a/src/coreclr/tools/Common/Internal/Runtime/ReadyToRunConstants.cs
+++ b/src/coreclr/tools/Common/Internal/Runtime/ReadyToRunConstants.cs
@@ -234,6 +234,7 @@ namespace Internal.ReadyToRunConstants
         WriteBarrier                = 0x30,
         CheckedWriteBarrier         = 0x31,
         ByRefWriteBarrier           = 0x32,
+        ByRefWriteBarrierBatch      = 0x33,
 
         // Array helpers
         Stelem_Ref                  = 0x38,