--- a/src/coreclr/tools/Common/JitInterface/CorInfoHelpFunc.cs
+++ b/src/coreclr/tools/Common/JitInterface/CorInfoHelpFunc.cs
@@ -138,6 +138,7 @@ namespace Internal.JitInterface
         CORINFO_HELP_ASSIGN_REF_ENSURE_NONHEAP,  // Do the store, and ensure that the target was not in the heap.
 
         CORINFO_HELP_ASSIGN_BYREF,
+        CORINFO_HELP_ASSIGN_BYREF_BATCH,
         CORINFO_HELP_ASSIGN_STRUCT,