--- a/src/coreclr/vm/jitinterface.h
+++ b/src/coreclr/vm/jitinterface.h
@@ -390,7 +390,7 @@ extern "C"
 
     void STDCALL JIT_ByRefWriteBarrier();      // JIThelp.asm/JIThelp.s
 
-#if defined(TARGET_AMD64) && defined(TARGET_WINDOWS)
+#if defined(TARGET_AMD64)
     void STDCALL JIT_ByRefWriteBarrierBatch(); // JIThelp.asm/JIThelp.s
 #endif // TARGET_AMD64 && TARGET_WINDOWS