… src/coreclr/inc/jithelpers.h
… src/coreclr/vm/excep.cpp