Closes https://github.com/dotnet/runtime/issues/8627

Testing on CI, for now - win-x64 coreclr only. Presumably, the asm version can be better than just a copy of `JIT_ByRefWriteBarrier` with a loop over R8 (length):

- [x] If first iteration finds out that the destination doesn't belong to any heap, we should skip all checks for the rest inside the batch
- [ ] TODO: what else
- [ ] Do we need debug-only `WRITE_BARRIER_CHECK`? we can use the non-batched version for DEBUG/CHECKED
- [ ] `REPRET` looks like some legacy we can replace with just a normal `ret`

[jit-diffs](https://gist.github.com/EgorBo/1e7dc8fecfe2ee0a53c5012101293da4) (seems like a few hits in Roslyn). Size regressions are from the cases when batchSize is 2, two calls are smaller than the batched one + `mov r8d, <size>`. But the perf should still be better with the batch one.

## Example:
```cs
struct MyStruct
{
    object a1;
    object a2;
    object a3;
    string a4;
    string a5;
    string a6;
}

static MyStruct Test(MyStruct ms) => ms; // simple copy of a struct with gc fields
```
#### Current codegen:
```asm
; Assembly listing for method Prog:Test(MyStruct):MyStruct
G_M32391_IG01:
       push     rdi
       push     rsi
       push     rbx
       mov      rbx, rcx
G_M32391_IG02:
       mov      rdi, rbx
       mov      rsi, rdx
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       call     CORINFO_HELP_ASSIGN_BYREF
       mov      rax, rbx
G_M32391_IG03:
       pop      rbx
       pop      rsi
       pop      rdi
       ret      
; Total bytes of code 49
```

#### New codegen:

```asm
; Assembly listing for method Prog:Test(MyStruct):MyStruct
G_M32391_IG01:
       push     rdi
       push     rsi
       push     rbx
       mov      rbx, rcx
G_M32391_IG02:
       mov      rdi, rbx
       mov      rsi, rdx
       mov      r8d, 6
       call     CORINFO_HELP_ASSIGN_BYREF_BATCH
       mov      rax, rbx
G_M32391_IG03:
       pop      rbx
       pop      rsi
       pop      rdi
       ret      
; Total bytes of code 30
```

## Benchmark:

```cs

struct MyStruct
{
    object a1;
    object a2;
    object a3;
    object a4;
}

[Benchmark]
public void StructCopy()
{
    MyStruct ms = default;
    Test(ms);

    [MethodImpl(MethodImplOptions.NoInlining)]
    MyStruct Test(MyStruct ms) => ms;
}
```

| Method     | Toolchain             | Mean     | Error     | StdDev    | Ratio |
|----------- |---------------------- |---------:|----------:|----------:|------:|
| StructCopy | Core_Root_Main\corerun.exe | 4.655 ns | 0.0097 ns | 0.0090 ns |  2.05 |
| StructCopy | Core_Root_PR\corerun.exe | **2.269 ns** | 0.0097 ns | 0.0086 ns |  1.00 |

When destination is on the heap, the perf is also improved