@EgorBo 2024-02-29
Benchmark:

```cs

struct MyStruct
{
    object a1;
    object a2;
    object a3;
    object a4;
}

[Benchmark]
public void StructCopy()
{
    MyStruct ms = default;
    Test(ms);

    [MethodImpl(MethodImplOptions.NoInlining)]
    MyStruct Test(MyStruct ms) => ms;
}
```

| Method     | Toolchain             | Mean     | Error     | StdDev    | Ratio |
|----------- |---------------------- |---------:|----------:|----------:|------:|
| StructCopy | Core_Root\corerun.exe | 4.655 ns | 0.0097 ns | 0.0090 ns |  2.05 |
| StructCopy | Core_Root\corerun.exe | **2.269 ns** | 0.0097 ns | 0.0086 ns |  1.00 |

When destination is on the heap, the perf is also improved