@jkotas 2024-02-29
Observation: The helper is functional equivalent of `BulkMoveWithWriteBarrier`, with different perf trade-off. `BulkMoveWithWriteBarrier`. Compared to`BulkMoveWithWriteBarrier` , the helper as currently implemented in the PR tries to be more precise with setting the cards so it is slower but leaves less work to be done at GC time.