@EgorBo 2024-02-29
> Observation: The helper is functional equivalent of `BulkMoveWithWriteBarrier`, with different perf trade-off. Compared to`BulkMoveWithWriteBarrier` , the helper as currently implemented in the PR tries to be more precise with setting the cards so it is slower but leaves less work to be done at GC time.

Thanks! Presumably it's easier to just call `BulkMoveWithWriteBarrier` so then we don't have to duplicate these helpers for all platforms. Do you have an opinion on which way is generally better?

One thing that motivated me is that we've seen write barriers on hot paths in several real world benchmarks (e.g. @kunalspathak recently did). I am not sure this specific one was the culprit, just looked like a low-hanging fruit to me.