@jkotas 2024-02-29
>  Presumably it's easier to just call BulkMoveWithWriteBarrier

BulkMoveWithWriteBarrier has higher fixed overhead than the hand-written assembly helper. It expect that it would be profitable to call it only for blocks above certain size.

> we've seen write barriers on hot paths in several real world benchmarks

Is this addressing the hot paths that we have seen?