@EgorBo 2024-02-29
> Is this addressing the hot paths that we have seen?

Not sure yet, but [jitdiffs](https://gist.github.com/EgorBo/1e7dc8fecfe2ee0a53c5012101293da4) imply they're not rare (jitdiffs are usually way smaller than SPMI because we only run it for framework libs, no tests, benchmarks, apsnet, etc).