@EgorBo 2024-02-29
Here is the current diff between non-batched (left) and batched (right) version: https://www.diffchecker.com/clp44lCx/