@VSadov 2024-03-01
My first thought about this is that the barriers are the kind of  platform- and achitecture-specific pieces of assembly, that we would avoid. With Core/AOT + 6 architecture + Win/Unix - how many combinations will need a copy of the new barrier?  Do we get enough savings for that?

It looks like it saves a call. Also hoists “not in heap” check, although that will help only if it is really not in heap and short-circuits, otherwise that check is cheap. The rest of the cost - other checks, lock or, the actual assignment are still there.
Considering the barrier cost is typically in 1-5% range and that this will be hit in fairly rare cases in the code (“few cases in Roslyn”?), I wonder how much of an improvement is expected here in general?

What happens to the benchmark if the write is to a bunch of array elements instead of a single stack location?