@EgorBo 2024-03-01
Closing, will try to benchmark/compare arm64 write barriers (not only byref one) against x64 to see if I can contribute there instead